MOving the tests to their right location
