+++ src/test/java/sa/com/cloudsolutions/antikythera/evaluator/TestIfConditionVisitor.java
@@ -1,0 +1,70 @@
+package sa.com.cloudsolutions.antikythera.evaluator;
+
+import com.github.javaparser.ast.CompilationUnit;
+import com.github.javaparser.ast.body.MethodDeclaration;
+import com.github.javaparser.ast.expr.Expression;
+import com.github.javaparser.ast.stmt.IfStmt;
+import com.github.javaparser.ast.visitor.VoidVisitorAdapter;
+import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Test;
+import sa.com.cloudsolutions.antikythera.configuration.Settings;
+import sa.com.cloudsolutions.antikythera.parser.AbstractCompiler;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.List;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+class TestIfConditionVisitor {
+    @BeforeAll
+    static void setup() throws IOException {
+        Settings.loadConfigMap(new File("src/test/resources/generator-field-tests.yml"));
+        AbstractCompiler.preProcess();
+    }
+
+    @Test
+    void testCollectConditionsUpToMethod() {
+        String cls = "sa.com.cloudsolutions.antikythera.evaluator.Conditional";
+        CompilationUnit cu = AntikytheraRunTime.getCompilationUnit(cls);
+        MethodDeclaration md = cu.findFirst(MethodDeclaration.class, f -> f.getNameAsString().equals("multiVariate")).get();
+        md.accept(new VoidVisitorAdapter<Void>() {
+                      @Override
+                      public void visit(IfStmt n, Void arg) {
+                          super.visit(n, arg);
+                          List<Expression> conditions =  IfConditionVisitor.collectConditionsUpToMethod(n);
+                          if (n.getCondition().toString().equals("a == 0")) {
+                              assertEquals(0, conditions.size());
+                          }
+                          else {
+                              assertEquals(1, conditions.size());
+                          }
+                      }
+                  }, null
+        );
+    }
+
+    @Test
+    void testCollectConditionsUpToMethodAgain() {
+        String cls = "sa.com.cloudsolutions.antikythera.evaluator.Conditional";
+        CompilationUnit cu = AntikytheraRunTime.getCompilationUnit(cls);
+        MethodDeclaration md = cu.findFirst(MethodDeclaration.class, f -> f.getNameAsString().equals("multiVariateDeep")).get();
+        md.accept(new VoidVisitorAdapter<Void>() {
+                      @Override
+                      public void visit(IfStmt n, Void arg) {
+                          super.visit(n, arg);
+                          List<Expression> conditions =  IfConditionVisitor.collectConditionsUpToMethod(n);
+                          if (n.getCondition().toString().equals("a == 0")) {
+                              assertEquals(0, conditions.size());
+                          }
+                          else if (n.getCondition().toString().equals("b == 1")) {
+                              assertEquals(2, conditions.size());
+                          }
+                          else {
+                              assertEquals(1, conditions.size());
+                          }
+                      }
+                  }, null
+        );
+    }
+}
