+++ src/test/java/sa/com/cloudsolutions/antikythera/evaluator/TestSpringEvaluator.java
@@ -3,12 +3,9 @@ package sa.com.cloudsolutions.antikythera.evaluator;
 import com.github.javaparser.StaticJavaParser;
 import com.github.javaparser.ast.body.FieldDeclaration;
 import com.github.javaparser.ast.body.VariableDeclarator;
-import com.github.javaparser.ast.expr.Expression;
 import com.github.javaparser.ast.expr.MethodCallExpr;
 import com.github.javaparser.ast.expr.NameExpr;
-import com.github.javaparser.ast.stmt.IfStmt;
 import com.github.javaparser.ast.type.ClassOrInterfaceType;
-import com.github.javaparser.ast.visitor.VoidVisitorAdapter;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -26,7 +23,6 @@ import com.github.javaparser.ast.body.MethodDeclaration;
 import java.io.File;
 import java.io.IOException;
 import java.lang.reflect.Field;
-import java.util.List;
 import java.util.NoSuchElementException;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -210,55 +206,6 @@ class TestSpringEvaluatorAgain {
     static void setup() throws IOException {
         Settings.loadConfigMap(new File("src/test/resources/generator-field-tests.yml"));
         AbstractCompiler.preProcess();
-    }
-
-    @Test
-    void testCollectConditionsUpToMethod() {
-        String cls = "sa.com.cloudsolutions.antikythera.evaluator.Conditional";
-        CompilationUnit cu = AntikytheraRunTime.getCompilationUnit(cls);
-        MethodDeclaration md = cu.findFirst(MethodDeclaration.class, f -> f.getNameAsString().equals("multiVariate")).get();
-        SpringEvaluator eval = EvaluatorFactory.create(cls, SpringEvaluator.class);
-
-        md.accept(new VoidVisitorAdapter<Void>() {
-                      @Override
-                      public void visit(IfStmt n, Void arg) {
-                          super.visit(n, arg);
-                          List<Expression> conditions =  IfConditionVisitor.collectConditionsUpToMethod(n);
-                          if (n.getCondition().toString().equals("a == 0")) {
-                              assertEquals(0, conditions.size());
-                          }
-                          else {
-                              assertEquals(1, conditions.size());
-                          }
-                      }
-                  }, null
-        );
-    }
-
-    @Test
-    void testCollectConditionsUpToMethodAgain() {
-        String cls = "sa.com.cloudsolutions.antikythera.evaluator.Conditional";
-        CompilationUnit cu = AntikytheraRunTime.getCompilationUnit(cls);
-        MethodDeclaration md = cu.findFirst(MethodDeclaration.class, f -> f.getNameAsString().equals("multiVariateDeep")).get();
-        SpringEvaluator eval = EvaluatorFactory.create(cls, SpringEvaluator.class);
-
-        md.accept(new VoidVisitorAdapter<Void>() {
-                      @Override
-                      public void visit(IfStmt n, Void arg) {
-                          super.visit(n, arg);
-                          List<Expression> conditions =  IfConditionVisitor.collectConditionsUpToMethod(n);
-                          if (n.getCondition().toString().equals("a == 0")) {
-                              assertEquals(0, conditions.size());
-                          }
-                          else if (n.getCondition().toString().equals("b == 1")) {
-                              assertEquals(2, conditions.size());
-                          }
-                          else {
-                              assertEquals(1, conditions.size());
-                          }
-                      }
-                  }, null
-        );
     }
 
     @Test
